# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CanteenIQ API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MessageResponse {
  message: string;
}

export type ProfileRole = (typeof ProfileRole)[keyof typeof ProfileRole];

export const ProfileRole = {
  student: "student",
  staff: "staff",
  admin: "admin",
} as const;

export interface Profile {
  id: number;
  fullName: string;
  email: string;
  role: ProfileRole;
  createdAt: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export type SignupBodyRole =
  (typeof SignupBodyRole)[keyof typeof SignupBodyRole];

export const SignupBodyRole = {
  student: "student",
  staff: "staff",
  admin: "admin",
} as const;

export interface SignupBody {
  email: string;
  password: string;
  fullName: string;
  role: SignupBodyRole;
}

export interface AuthResponse {
  user: Profile;
  token: string;
}

export type MenuItemVegType =
  (typeof MenuItemVegType)[keyof typeof MenuItemVegType];

export const MenuItemVegType = {
  veg: "veg",
  nonveg: "nonveg",
  egg: "egg",
} as const;

export interface MenuItem {
  id: number;
  name: string;
  category: string;
  price: number;
  imageUrl?: string;
  available: boolean;
  popular: boolean;
  vegType: MenuItemVegType;
  stock: number;
  createdAt: string;
}

export type CreateMenuItemBodyVegType =
  (typeof CreateMenuItemBodyVegType)[keyof typeof CreateMenuItemBodyVegType];

export const CreateMenuItemBodyVegType = {
  veg: "veg",
  nonveg: "nonveg",
  egg: "egg",
} as const;

export interface CreateMenuItemBody {
  name: string;
  category: string;
  price: number;
  imageUrl?: string;
  available: boolean;
  popular: boolean;
  vegType: CreateMenuItemBodyVegType;
  stock: number;
}

export type UpdateMenuItemBodyVegType =
  (typeof UpdateMenuItemBodyVegType)[keyof typeof UpdateMenuItemBodyVegType];

export const UpdateMenuItemBodyVegType = {
  veg: "veg",
  nonveg: "nonveg",
  egg: "egg",
} as const;

export interface UpdateMenuItemBody {
  name?: string;
  category?: string;
  price?: number;
  imageUrl?: string;
  available?: boolean;
  popular?: boolean;
  vegType?: UpdateMenuItemBodyVegType;
  stock?: number;
}

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  pending: "pending",
  preparing: "preparing",
  ready: "ready",
  picked_up: "picked_up",
  cancelled: "cancelled",
} as const;

export interface Order {
  id: number;
  userId: number;
  tokenNumber: number;
  pickupTime: string;
  status: OrderStatus;
  totalAmount: number;
  createdAt: string;
  userFullName?: string;
  userEmail?: string;
}

export interface OrderItem {
  id: number;
  orderId: number;
  menuItemId: number;
  quantity: number;
  price: number;
  menuItemName?: string;
}

export type OrderDetailStatus =
  (typeof OrderDetailStatus)[keyof typeof OrderDetailStatus];

export const OrderDetailStatus = {
  pending: "pending",
  preparing: "preparing",
  ready: "ready",
  picked_up: "picked_up",
  cancelled: "cancelled",
} as const;

export interface OrderDetail {
  id: number;
  userId: number;
  tokenNumber: number;
  pickupTime: string;
  status: OrderDetailStatus;
  totalAmount: number;
  createdAt: string;
  userFullName?: string;
  items: OrderItem[];
}

export type CreateOrderBodyItemsItem = {
  menuItemId: number;
  quantity: number;
};

export interface CreateOrderBody {
  pickupTime: string;
  items: CreateOrderBodyItemsItem[];
}

export type UpdateOrderStatusBodyStatus =
  (typeof UpdateOrderStatusBodyStatus)[keyof typeof UpdateOrderStatusBodyStatus];

export const UpdateOrderStatusBodyStatus = {
  pending: "pending",
  preparing: "preparing",
  ready: "ready",
  picked_up: "picked_up",
  cancelled: "cancelled",
} as const;

export interface UpdateOrderStatusBody {
  status: UpdateOrderStatusBodyStatus;
}

export type PredictionWasteRisk =
  (typeof PredictionWasteRisk)[keyof typeof PredictionWasteRisk];

export const PredictionWasteRisk = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export type PredictionTrend =
  (typeof PredictionTrend)[keyof typeof PredictionTrend];

export const PredictionTrend = {
  up: "up",
  down: "down",
  stable: "stable",
} as const;

export interface Prediction {
  id: number;
  itemName: string;
  predictedQuantity: number;
  confidence: number;
  wasteRisk: PredictionWasteRisk;
  yesterdaySold: number;
  trend: PredictionTrend;
  date: string;
}

export interface Feedback {
  id: number;
  userId: number;
  rating: number;
  message: string;
  createdAt: string;
  userFullName?: string;
}

export interface SubmitFeedbackBody {
  rating: number;
  message: string;
}

export interface SupportTicket {
  id: number;
  userId: number;
  issueType: string;
  message: string;
  status: string;
  createdAt: string;
  userFullName?: string;
}

export interface CreateSupportTicketBody {
  issueType: string;
  message: string;
}

export interface DashboardSummary {
  totalOrdersToday: number;
  revenueToday: number;
  wasteEstimate: number;
  topSellingItem: string;
  peakHour: string;
  pendingOrders: number;
  preparingOrders: number;
  readyOrders: number;
  completedOrders: number;
  cancelledOrders: number;
}

export interface SalesDataPoint {
  label: string;
  revenue: number;
  orders: number;
}

export interface CategorySales {
  category: string;
  revenue: number;
  count: number;
}

export interface TopItem {
  name: string;
  count: number;
  revenue: number;
}

export interface SalesAnalytics {
  data: SalesDataPoint[];
  categoryBreakdown: CategorySales[];
  topItems: TopItem[];
}

export interface PeakHourData {
  hour: string;
  day: string;
  orders: number;
}

export interface SustainabilityMetrics {
  wasteReducedPercent: number;
  foodSavedKg: number;
  sustainabilityScore: number;
  carbonImpactKg: number;
  improvedDemandPlanningPercent: number;
  monthlyFoodSaved: number;
}

export interface StaffDailySummary {
  totalOrdersServed: number;
  mostOrderedItem: string;
  cancelledOrders: number;
  pendingOrders: number;
  busiestHour: string;
  totalRevenue: number;
  quantitySold: number;
}

export type ListMenuItemsParams = {
  category?: string;
  available?: boolean;
  search?: string;
};

export type ListOrdersParams = {
  status?: ListOrdersStatus;
  date?: string;
};

export type ListOrdersStatus =
  (typeof ListOrdersStatus)[keyof typeof ListOrdersStatus];

export const ListOrdersStatus = {
  pending: "pending",
  preparing: "preparing",
  ready: "ready",
  picked_up: "picked_up",
  cancelled: "cancelled",
} as const;

export type ListPredictionsParams = {
  date?: string;
};

export type GetSalesAnalyticsParams = {
  period?: GetSalesAnalyticsPeriod;
};

export type GetSalesAnalyticsPeriod =
  (typeof GetSalesAnalyticsPeriod)[keyof typeof GetSalesAnalyticsPeriod];

export const GetSalesAnalyticsPeriod = {
  daily: "daily",
  weekly: "weekly",
  monthly: "monthly",
} as const;