# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CanteenIQ API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user profile
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "401":
          description: Unauthorized
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
  /auth/signup:
    post:
      operationId: signup
      tags: [auth]
      summary: Sign up with email, password, name, and role
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupBody"
      responses:
        "201":
          description: Signup successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Email already in use
...
  /menu:
    get:
      operationId: listMenuItems
      tags: [menu]
      summary: List all menu items
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: available
          in: query
          schema:
            type: boolean
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of menu items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MenuItem"
    post:
      operationId: createMenuItem
      tags: [menu]
      summary: Create a new menu item (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMenuItemBody"
      responses:
        "201":
          description: Created menu item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
  /menu/{id}:
    get:
      operationId: getMenuItem
      tags: [menu]
      summary: Get a single menu item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Menu item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
    put:
      operationId: updateMenuItem
      tags: [menu]
      summary: Update a menu item (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMenuItemBody"
      responses:
        "200":
          description: Updated menu item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
    delete:
      operationId: deleteMenuItem
      tags: [menu]
      summary: Delete a menu item (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders (students see own, staff/admin see all)
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, preparing, ready, picked_up, cancelled]
        - name: date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place a new order (student)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderBody"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get a single order with items
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderDetail"
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status (staff/admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusBody"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}/cancel:
    post:
      operationId: cancelOrder
      tags: [orders]
      summary: Cancel an order (student - only pending orders)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Cancelled order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    Profile:
      type: object
      properties:
        id:
          type: integer
        fullName:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [student, staff, admin]
        createdAt:
          type: string
      required:
        - id
        - fullName
        - email
        - role
        - createdAt
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    SignupBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        fullName:
          type: string
        role:
          type: string
          enum: [student, staff, admin]
      required:
        - email
        - password
        - fullName
        - role
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/Profile"
        token:
          type: string
      required:
        - user
        - token
    MenuItem:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        price:
          type: number
        imageUrl:
          type: string
        available:
          type: boolean
        popular:
          type: boolean
        vegType:
          type: string
          enum: [veg, nonveg, egg]
        stock:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - name
        - category
        - price
        - available
        - popular
        - vegType
        - stock
        - createdAt
    CreateMenuItemBody:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        price:
          type: number
        imageUrl:
          type: string
        available:
          type: boolean
        popular:
          type: boolean
        vegType:
          type: string
          enum: [veg, nonveg, egg]
        stock:
          type: integer
      required:
        - name
        - category
        - price
        - available
        - popular
        - vegType
        - stock
    UpdateMenuItemBody:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        price:
          type: number
        imageUrl:
          type: string
        available:
          type: boolean
        popular:
          type: boolean
        vegType:
          type: string
          enum: [veg, nonveg, egg]
        stock:
          type: integer
    Order:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        tokenNumber:
          type: integer
        pickupTime:
          type: string
        status:
          type: string
          enum: [pending, preparing, ready, picked_up, cancelled]
        totalAmount:
          type: number
        createdAt:
          type: string
        userFullName:
          type: string
        userEmail:
          type: string
      required:
        - id
        - userId
        - tokenNumber
        - pickupTime
        - status
        - totalAmount
        - createdAt
    OrderItem:
      type: object
      properties:
        id:
          type: integer
        orderId:
          type: integer
        menuItemId:
          type: integer
        quantity:
          type: integer
        price:
          type: number
        menuItemName:
          type: string
      required:
        - id
        - orderId
        - menuItemId
        - quantity
        - price
    OrderDetail:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        tokenNumber:
          type: integer
        pickupTime:
          type: string
        status:
          type: string
          enum: [pending, preparing, ready, picked_up, cancelled]
        totalAmount:
          type: number
        createdAt:
          type: string
        userFullName:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
      required:
        - id
        - userId
        - tokenNumber
        - pickupTime
        - status
        - totalAmount
        - createdAt
        - items
    CreateOrderBody:
      type: object
      properties:
        pickupTime:
          type: string
        items:
          type: array
          items:
            type: object
            properties:
              menuItemId:
                type: integer
              quantity:
                type: integer
            required:
              - menuItemId
              - quantity
      required:
        - pickupTime
        - items
    UpdateOrderStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, preparing, ready, picked_up, cancelled]
      required:
        - status
...