# lib/api-zod/src/generated/types/createMenuItemBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CanteenIQ API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateMenuItemBodyVegType } from "./createMenuItemBodyVegType";

export interface CreateMenuItemBody {
  name: string;
  category: string;
  price: number;
  imageUrl?: string;
  available: boolean;
  popular: boolean;
  vegType: CreateMenuItemBodyVegType;
  stock: number;
}