# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CanteenIQ API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderStatus } from "./orderStatus";

export interface Order {
  id: number;
  userId: number;
  tokenNumber: number;
  pickupTime: string;
  status: OrderStatus;
  totalAmount: number;
  createdAt: string;
  userFullName?: string;
  userEmail?: string;
}