# lib/api-zod/src/generated/types/signupBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CanteenIQ API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SignupBodyRole } from "./signupBodyRole";

export interface SignupBody {
  email: string;
  password: string;
  fullName: string;
  role: SignupBodyRole;
}